refactor(web): migrate lib/uri to ReScript (migration Phase 0)
First real ReScript module, proving the interop loop end-to-end:

- Port lib/uri.ts -> lib/Uri.res; genType emits Uri.gen.tsx with the exact
  same TS signature (string | null | undefined) => string, so the three
  genre-page callers just switch their import to ./Uri.gen. Compiled JS is
  behavior-identical, including the defensive null/malformed-uri -> "#" path.
- Enable genType via rescript.json gentypeconfig (.gen.tsx output).
- Add src/rescript.d.ts ambient decl so the generated files' `.res.mjs`
  imports resolve under strict/noImplicitAny.
- Regenerate ReScript output on dev/build via pre{dev,prod,build,build:prod}
  hooks; gitignore the generated *.res.mjs / *.gen.tsx and the lib/ cache.
- Fix the gitignore `lib/` rule that also matched src/lib (anchor to /lib/),
  which would have silently made any new src/lib file un-committable.
- Drop the Test.res scaffolding.

Verified: rescript build, tsc -b, and full vite build all pass.
