+++ apps/web/.gitignore
@@ -37,5 +37,10 @@ dist-ssr
 # rockbox-wasm runtime assets (copied from node_modules by vite.config.ts)
 public/rockbox
 
-# ReScript build cache
-lib/
+# ReScript build cache (anchored to apps/web/lib — must NOT match src/lib source)
+/lib/
+watch.lock
+.sourcedirs.json
+# ReScript generated output (regenerated by `rescript build`; see pre{dev,build})
+*.res.mjs
+*.gen.tsx
