+++ apps/web/package.json
@@ -7,9 +7,13 @@
     "res:build": "rescript build",
     "res:dev": "rescript watch",
     "res:clean": "rescript clean",
+    "predev": "rescript build",
     "dev": "vite --port 5174",
+    "preprod": "rescript build",
     "prod": "vite --mode prod",
+    "prebuild": "rescript build",
     "build": "tsc -b && vite build",
+    "prebuild:prod": "rescript build",
     "build:prod": "tsc -b && vite build --mode prod",
     "lint": "biome lint src",
     "preview": "bun run build && wrangler pages dev ./dist",
