--- apps/web/src/Test.res
@@ -1,4 +1,0 @@
-@react.component
-let make = () => {
-  <div> {React.string("Hello World")} </div>
-}
