--- apps/web/src/Test.res.mjs
@@ -1,16 +1,0 @@
-// Generated by ReScript, PLEASE EDIT WITH CARE
-
-import * as JsxRuntime from "react/jsx-runtime";
-
-function Test(props) {
-  return JsxRuntime.jsx("div", {
-    children: "Hello World"
-  });
-}
-
-let make = Test;
-
-export {
-  make,
-}
-/* react/jsx-runtime Not a pure module */
